+++ scripts/sql_s28_batch.sql
@@ -1,0 +1,88 @@
+BEGIN;
+
+-- Moundville (id=775) - 1050-1500 Mississippian
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(775, 'Blitz, J. H. (2008). Moundville. University of Alabama Press. ISBN 978-0817354787.', 'https://www.uapress.ua.edu/9780817354787/moundville/', 'academic'),
+(775, 'Knight, V. J. Jr. (2010). Mound Excavations at Moundville: Architecture, Elites, and Social Order. University of Alabama Press. ISBN 978-0817316877.', 'https://www.uapress.ua.edu/9780817316877/', 'academic'),
+(775, 'Steponaitis, V. P. (1983). Ceramics, Chronology, and Community Patterns: An Archaeological Study at Moundville. Academic Press / University of Alabama Press reprint. ISBN 978-0817351809.', 'https://www.uapress.ua.edu/9780817351809/', 'academic'),
+(775, 'Wilson, G. D. (2008). The Archaeology of Everyday Life at Early Moundville. University of Alabama Press. ISBN 978-0817354442.', 'https://www.uapress.ua.edu/9780817354442/', 'academic'),
+(775, 'Steponaitis, V. P. & Scarry, C. M. (eds.) (2016). Rethinking Moundville and Its Hinterland. University Press of Florida. ISBN 978-0813061665.', 'https://upf.com/book.asp?id=STEPO001', 'academic');
+
+-- Electorate of Saxony (id=580) - 1356-1806
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(580, 'Wilson, P. H. (2016). Heart of Europe: A History of the Holy Roman Empire. Belknap Press of Harvard University Press. ISBN 978-0674058095.', 'https://www.hup.harvard.edu/catalog.php?isbn=9780674058095', 'academic'),
+(580, 'Groß, R. (2007). Die Wettiner. Kohlhammer Verlag. ISBN 978-3170189461.', 'https://www.kohlhammer.de/wms/instances/KOB/appDE/Geschichte/Allgemeine-Geschichte/Die-Wettiner/', 'academic'),
+(580, 'Kroll, F.-L. (ed.) (2007). Die Herrscher Sachsens: Markgrafen, Kurfürsten, Könige 1089-1918. C.H. Beck. ISBN 978-3406547737.', 'https://www.chbeck.de/9783406547737', 'academic'),
+(580, 'Blaschke, K. (2002). Die Leipziger Teilung der wettinischen Länder von 1485. In: Beiträge zur Verfassungs- und Verwaltungsgeschichte Sachsens, pp. 323-335. Leipziger Universitätsverlag. ISBN 978-3936522020.', 'https://www.univerlag-leipzig.de', 'academic'),
+(580, 'Schirmer, U. (2006). Kursächsische Staatsfinanzen 1456-1656. Franz Steiner Verlag. ISBN 978-3515089555.', 'https://www.steiner-verlag.de', 'academic');
+
+-- Württemberg (id=587) - 1495-1806
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(587, 'Wilson, P. H. (2016). Heart of Europe: A History of the Holy Roman Empire. Belknap Press of Harvard University Press. ISBN 978-0674058095.', 'https://www.hup.harvard.edu/catalog.php?isbn=9780674058095', 'academic'),
+(587, 'Marcus, K. H. (2000). The Politics of Power: Elites of an Early Modern State in Germany. Verlag Philipp von Zabern. ISBN 978-3805325349.', 'https://www.zabern.de', 'academic'),
+(587, 'Fulbrook, M. (1983). Piety and Politics: Religion and the Rise of Absolutism in England, Württemberg, and Prussia. Cambridge University Press. ISBN 978-0521256124.', 'https://www.cambridge.org/9780521256124', 'academic'),
+(587, 'Raitt, J. (1993). The Colloquy of Montbéliard: Religion and Politics in the Sixteenth Century. Oxford University Press. ISBN 978-0195075663.', 'https://global.oup.com/academic/product/the-colloquy-of-montbeliard-9780195075663', 'academic'),
+(587, 'Vann, J. A. (1984). The Making of a State: Württemberg, 1593-1793. Cornell University Press. ISBN 978-0801415029.', 'https://www.cornellpress.cornell.edu/book/9780801415029/', 'academic');
+
+-- Damot (id=832) - 1100-1316 medieval Ethiopia
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(832, 'Bouanga, A. (2014). Le royaume du Damot: enquête sur une puissance politique et économique de la Corne de l''Afrique (XIIIe siècle). Annales d''Éthiopie, 29, 27-58. DOI:10.3406/ethio.2014.1572.', 'https://doi.org/10.3406/ethio.2014.1572', 'academic'),
+(832, 'Kelly, S. (ed.) (2020). A Companion to Medieval Ethiopia and Eritrea. Brill. ISBN 978-9004419605.', 'https://brill.com/view/title/38063', 'academic'),
+(832, 'Hassen, M. (2015). The Oromo and the Christian Kingdom of Ethiopia: 1300-1700. James Currey / Boydell & Brewer. ISBN 978-1847011176.', 'https://www.jamescurrey.com/9781847011176', 'academic'),
+(832, 'Shinn, D. H. & Ofcansky, T. P. (2013). Historical Dictionary of Ethiopia (2nd ed.). Scarecrow Press. ISBN 978-0810874572.', 'https://rowman.com/ISBN/9780810874572', 'academic'),
+(832, 'Taddesse Tamrat (1972). Church and State in Ethiopia, 1270-1527. Oxford University Press / Clarendon Press. ISBN 978-0198216711.', 'https://global.oup.com/academic/product/church-and-state-in-ethiopia-9780198216711', 'academic');
+
+-- Bigo bya Mugenyi (id=996) - 1300-1500 Bunyoro precursor
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(996, 'Sutton, J. E. G. (1998). Ntusi and Bigo: Farmers, cattle-herders and rulers in western Uganda, AD 1000-1500. Azania, 33, 39-72. DOI:10.1080/00672709809511415.', 'https://doi.org/10.1080/00672709809511415', 'academic'),
+(996, 'Posnansky, M. (1969). Bigo bya Mugenyi. The Uganda Journal, 33(2), 125-150.', 'https://digital.lib.msu.edu/projects/africanjournals', 'academic'),
+(996, 'Robertshaw, P. (2001). Ancient Earthworks of Western Uganda: Capital Sites of a Cwezi Empire? The Uganda Journal, 47, 20-33.', 'https://digital.lib.msu.edu/projects/africanjournals', 'academic'),
+(996, 'Reid, A. (1996). Cattle Herds and the Redistribution of Cattle Resources. World Archaeology, 28(1), 43-57. DOI:10.1080/00438243.1996.9980329.', 'https://doi.org/10.1080/00438243.1996.9980329', 'academic'),
+(996, 'Schoenbrun, D. L. (1998). A Green Place, A Good Place: Agrarian Change, Gender, and Social Identity in the Great Lakes Region to the 15th Century. James Currey / Heinemann. ISBN 978-0852557341.', 'https://www.jamescurrey.com/9780852557341', 'academic');
+
+-- K2 / Bambandyanalo (id=990) - 1000-1075 Limpopo
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(990, 'Huffman, T. N. (2005). Mapungubwe: Ancient African Civilisation on the Limpopo. Wits University Press. ISBN 978-1868144082.', 'https://witspress.co.za/catalogue/mapungubwe/', 'academic'),
+(990, 'Calabrese, J. A. (2000). Interregional Interaction in Southern Africa: Zhizo and Leopard''s Kopje Relations in Northern South Africa, Southwestern Zimbabwe, and Eastern Botswana, AD 1000-1200. African Archaeological Review, 17(4), 183-210. DOI:10.1023/A:1006796925891.', 'https://doi.org/10.1023/A:1006796925891', 'academic'),
+(990, 'Chirikure, S., Manyanga, M., Pollard, A. M., Bandama, F., Mahachi, G. & Pikirayi, I. (2014). Zimbabwe Culture before Mapungubwe: New Evidence from Mapela Hill, South-Western Zimbabwe. PLOS ONE, 9(10), e111224. DOI:10.1371/journal.pone.0111224.', 'https://doi.org/10.1371/journal.pone.0111224', 'academic'),
+(990, 'Meyer, A. (1998). The Archaeological Sites of Greefswald: Stratigraphy and Chronology of the Sites and a History of Investigations. University of Pretoria. ISBN 978-0869801826.', 'https://www.up.ac.za', 'academic'),
+(990, 'Kim, N. C. & Kusimba, C. M. (2008). Pathways to Social Complexity and State Formation in the Southern Zambezian Region. African Archaeological Review, 25(3-4), 131-152. DOI:10.1007/s10437-008-9036-y.', 'https://doi.org/10.1007/s10437-008-9036-y', 'academic');
+
+-- Chanka (id=952) - 1000-1440 Andean
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(952, 'Rostworowski de Diez Canseco, M. (1999). History of the Inca Realm. Cambridge University Press. ISBN 978-0521637596.', 'https://www.cambridge.org/9780521637596', 'academic'),
+(952, 'Bauer, B. S., Kellett, L. C. & Aráoz Silva, M. (2010). The Chanka: Archaeological Research in Andahuaylas (Apurímac), Peru. Cotsen Institute of Archaeology Press / UCLA. ISBN 978-1931745543.', 'https://ioa.ucla.edu/cotsen-press/the-chanka', 'academic'),
+(952, 'Kellett, L. C. (2010). A Provincial Perspective of Wari and Chanka Settlement Patterns in Andahuaylas, Peru. PhD dissertation, University of New Mexico.', 'https://digitalrepository.unm.edu/anth_etds/86/', 'academic'),
+(952, 'D''Altroy, T. N. (2014). The Incas (2nd ed.). Wiley-Blackwell. ISBN 978-1444331158.', 'https://www.wiley.com/en-us/The+Incas%2C+2nd+Edition-p-9781444331158', 'academic'),
+(952, 'Meddens, F. (1985). The Chanca Confederation: Political Myth and Archaeological Reality. Papers from the Institute of Archaeology, 16, 73-99.', 'https://www.ucl.ac.uk/archaeology/pia', 'academic');
+
+-- Bono-Manso (id=1014) - 1270-1500 Akan precursor
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1014, 'Effah-Gyamfi, K. (1985). Bono Manso: An Archaeological Investigation into Early Akan Urbanism. African Occasional Papers 2, Department of Archaeology, University of Calgary. ISBN 978-0919813328.', 'https://www.ucalgary.ca', 'academic'),
+(1014, 'Warren, D. M. (1976). Bono-Manso Chronology: A Methodological Reassessment. Journal of African History, 17(3), 366-377. DOI:10.1017/S0021853700014857.', 'https://doi.org/10.1017/S0021853700014857', 'academic'),
+(1014, 'Konadu, K. (2010). The Akan Diaspora in the Americas. Oxford University Press. ISBN 978-0195390643.', 'https://global.oup.com/academic/product/the-akan-diaspora-in-the-americas-9780195390643', 'academic'),
+(1014, 'Posnansky, M. (2015). Begho: Life and Times. Journal des Africanistes, 85(1-2), 280-299.', 'https://www.persee.fr/issue/jafr_0399-0346_2015_num_85_1', 'academic'),
+(1014, 'Wilks, I. (1993). Forests of Gold: Essays on the Akan and the Kingdom of Asante. Ohio University Press. ISBN 978-0821410288.', 'https://www.ohioswallow.com/book/Forests+of+Gold', 'academic');
+
+-- Kong Empire (id=1013) - 1300-1500 Niger Bend
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1013, 'Şaul, M. (1998). The War Houses of the Watara in West Africa. International Journal of African Historical Studies, 31(3), 537-570. DOI:10.2307/220977.', 'https://doi.org/10.2307/220977', 'academic'),
+(1013, 'Tauxier, L. & Bernus, E. (eds.) (2003). Les états de Kong. Karthala Editions. ISBN 978-2845863156.', 'https://www.karthala.com', 'academic'),
+(1013, 'Azarya, V. (1988). Jihads and the Dyula State in West Africa. In: S. N. Eisenstadt et al. (eds.), The Early State in African Perspective, pp. 137-170. Brill. ISBN 978-9004085770.', 'https://brill.com/view/title/4076', 'academic'),
+(1013, 'Launay, R. (1988). Warriors and Traders: The Political Organization of a West African Chiefdom. Cahiers d''Études Africaines, 28(111-112), 355-373.', 'https://www.persee.fr/issue/cea_0008-0055_1988_num_28_111', 'academic'),
+(1013, 'Massing, A. W. (2000). The Wangara, an Old Soninke Diaspora in West Africa? Cahiers d''Études Africaines, 40(158), 281-308.', 'https://www.persee.fr/issue/cea_0008-0055_2000_num_40_158', 'academic');
+
+-- Lenapehoking (id=779) - 1000-1763
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(779, 'Weslager, C. A. (1972). The Delaware Indians: A History. Rutgers University Press. ISBN 978-0813503929.', 'https://www.rutgersuniversitypress.org/the-delaware-indians/9780813503929', 'academic'),
+(779, 'Goddard, I. (1978). Delaware. In: B. G. Trigger (ed.), Handbook of North American Indians, Vol. 15: Northeast, pp. 213-239. Smithsonian Institution. ISBN 978-0160045752.', 'https://www.bookstore.gpo.gov', 'academic'),
+(779, 'Schutt, A. C. (2007). Peoples of the River Valleys: The Odyssey of the Delaware Indians. University of Pennsylvania Press. ISBN 978-0812240313.', 'https://www.upenn.edu/pennpress/book/14367.html', 'academic'),
+(779, 'Newcomb, W. W. Jr. (1956). The Culture and Acculturation of the Delaware Indians. Anthropological Papers, Museum of Anthropology, University of Michigan, No. 10.', 'https://www.lsa.umich.edu/ummaa', 'academic'),
+(779, 'Grumet, R. S. (2009). The Munsee Indians: A History. University of Oklahoma Press. ISBN 978-0806139838.', 'https://www.oupress.com/9780806139838/the-munsee-indians', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (775, 580, 587, 832, 996, 990, 952, 1014, 1013, 779);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (580, 587, 775, 779, 832, 952, 990, 996, 1013, 1014) ORDER BY g.id;